requirements.txt
# Core dependencies for SoundHire Cloud application
supabase>=2.0.0
python-dotenv==1.0.0